+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.0.2
+
+- Added minimum density of wedges.
+  - When total wedge count is low, weights are automatically "upscaled".
+  - This helps keep lots of small wedges even if only using a few options.
+- Added settings modal.
+  - Added spin duration setting.
+  - Added minimum density settings.
+- Bugfix: Color picker will close if open when add or delete occurs.
+
 # 1.0.1
 
 - Overhauled themes. Changed the technical execution to then allow making more pleasing aesthetic choices.